# apps/docs/public/platform-api.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VhyxVoid platform API",
    "version": "1.0.0",
    "description": "The dashboard's API, for scripts, CI and editors. Authenticate with an API key: `Authorization: Bearer <keyId>.<secret>`. A key acts in its own workspace, as the member who created it, and each route needs one scope. Each key may make a number of requests per minute (plan limit); every response carries X-RateLimit headers. Generated from the routes the API serves."
  },
  "servers": [
    {
      "url": "https://api.vhyxvoid.com"
    }
  ],
...
  "paths": {
...
    "/api/v1/team/{accountId}/comments": {
      "get": {
        "operationId": "getTeamByAccountIdComments",
        "tags": [
          "Team space"
        ],
        "summary": "List comments",
        "description": "On a document (target doc:<id>) or an issue (target issue:<number>).\n\nNeeds the `team:read` scope.",
        "x-required-scope": "team:read",
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "target",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^(doc:[0-9a-f-]{36}|issue:\\d{1,9})$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "headers": {
              "X-RateLimit-Limit": {
                "$ref": "#/components/headers/X-RateLimit-Limit"
              },
              "X-RateLimit-Remaining": {
                "$ref": "#/components/headers/X-RateLimit-Remaining"
              },
              "X-RateLimit-Reset": {
                "$ref": "#/components/headers/X-RateLimit-Reset"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/E400"
          },
          "401": {
            "$ref": "#/components/responses/E401"
          },
          "402": {
            "$ref": "#/components/responses/E402"
          },
          "403": {
            "$ref": "#/components/responses/E403"
          },
          "404": {
            "$ref": "#/components/responses/E404"
          },
          "429": {
            "$ref": "#/components/responses/E429"
          }
        }
      },
      "post": {
        "operationId": "postTeamByAccountIdComments",
        "tags": [
          "Team space"
        ],
        "summary": "Comment",
        "description": "On a document (anchor: a heading id) or an issue. Notifies mentions, and the document's author or the issue's assignee and creator.\n\nNeeds the `team:write` scope.",
        "x-required-scope": "team:write",
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "target": {
                    "type": "string",
                    "pattern": "^(doc:[0-9a-f-]{36}|issue:\\d{1,9})$"
                  },
                  "body": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 8000
                  },
                  "anchor": {
                    "anyOf": [
                      {
                        "type": "string",
                        "maxLength": 120
                      },
                      {
                        "type": "null"
                      }
                    ]
                  }
                },
                "required": [
                  "target",
                  "body"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success.",
            "headers": {
              "X-RateLimit-Limit": {
                "$ref": "#/components/headers/X-RateLimit-Limit"
              },
              "X-RateLimit-Remaining": {
                "$ref": "#/components/headers/X-RateLimit-Remaining"
              },
              "X-RateLimit-Reset": {
                "$ref": "#/components/headers/X-RateLimit-Reset"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/E400"
          },
          "401": {
            "$ref": "#/components/responses/E401"
          },
          "402": {
            "$ref": "#/components/responses/E402"
          },
          "403": {
            "$ref": "#/components/responses/E403"
          },
          "404": {
            "$ref": "#/components/responses/E404"
          },
          "429": {
            "$ref": "#/components/responses/E429"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "keyId.secret",
        "description": "An API key from the dashboard: the key ID, a dot, and the secret."
      }
    },
    "headers": {
      "X-RateLimit-Limit": {
        "description": "The key's requests per minute.",
        "schema": {
          "type": "integer"
        }
      },
      "X-RateLimit-Remaining": {
        "description": "Requests left this minute.",
        "schema": {
          "type": "integer"
        }
      },
      "X-RateLimit-Reset": {
        "description": "Seconds until the minute ends.",
        "schema": {
          "type": "integer"
        }
      },
      "Retry-After": {
        "description": "Seconds to wait before trying again.",
        "schema": {
          "type": "integer"
        }
      }
    },
    "schemas": {
      "Success": {
        "type": "object",
        "required": [
          "success",
          "message",
          "data"
        ],
        "properties": {
          "success": {
            "const": true
          },
          "message": {
            "type": "string"
          },
          "data": {
            "description": "The result; its shape is in the route's description and the docs."
          }
        }
      },
      "Error": {
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "properties": {
          "success": {
            "const": false
          },
          "message": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "data": {
            "type": "null"
          },
          "requestId": {
            "type": "string"
          }
        }
      }
    },
    "responses": {
      "E400": {
        "description": "The request is not valid; the message says why.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "E401": {
        "description": "The key is missing, wrong, revoked or expired, or this route needs a dashboard sign-in.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "E402": {
        "description": "A plan limit was reached.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "E403": {
        "description": "The key lacks the route's scope, belongs to another workspace, or its creator is no longer a member.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "E404": {
        "description": "Not found.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "E429": {
        "description": "The key's requests-per-minute limit (or the per-address limit) was reached. Retry after Retry-After.",
        "headers": {
          "Retry-After": {
            "$ref": "#/components/headers/Retry-After"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  }
}